% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georeference.R
\name{build_virtual_raster}
\alias{build_virtual_raster}
\title{Build a virtual raster from GeoTIFFs}
\usage{
build_virtual_raster(tile_numbers_df, vrt_file)
}
\description{
Build a virtual raster from GeoTIFFs
}
