% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_index.R
\name{corner_coords}
\alias{corner_coords}
\title{Find coordinates of the upper left and lower right corners of a tile}
\usage{
corner_coords(zoom, xtile, ytile)
}
\description{
Find coordinates of the upper left and lower right corners of a tile
}
