% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbox.R
\name{download_mapbox_tiles}
\alias{download_mapbox_tiles}
\title{Download tiles from mapbox given zoom level and the calculated tile indices}
\usage{
download_mapbox_tiles(
  tile_numbers_mat,
  download_dir,
  resolution = "high",
  jpg_quality = 90
)
}
\description{
Download tiles from mapbox given zoom level and the calculated tile indices
}
