% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_index.R
\name{find_tile_number}
\alias{find_tile_number}
\title{Find all tiles given a zoom level and the coordinates of each corner}
\usage{
find_tile_number(zoom, lat_deg, lon_deg)
}
\description{
Find all tiles given a zoom level and the coordinates of each corner
}
