% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_index.R
\name{find_tile_numbers}
\alias{find_tile_numbers}
\title{Find all tiles given a zoom level and the coordinates of each corner}
\usage{
find_tile_numbers(zoom, upper_left, lower_right)
}
\description{
Find all tiles given a zoom level and the coordinates of each corner
}
