% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georeference.R
\name{georeference_all_tiles}
\alias{georeference_all_tiles}
\title{Walk through tile file names and georeference all of them}
\usage{
georeference_all_tiles(tile_numbers_df)
}
\description{
Walk through tile file names and georeference all of them
}
