% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georeference.R
\name{georeference_tile}
\alias{georeference_tile}
\title{Georeference a Mapbox jpg tile and create GeoTiff}
\usage{
georeference_tile(file_name, output_file_name, x, y, zoom, ...)
}
\description{
Georeference a Mapbox jpg tile and create GeoTiff
}
